requirements.txt
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.0
asyncpg==0.30.0
click==8.3.0
colorama==0.4.6
fastapi==0.119.0
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
pydantic==2.12.1
pydantic_core==2.41.3
python-dotenv==1.1.1
PyYAML==6.0.3
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
